+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `wal-upgrade` to call `package-vc-upgrade-all` for `wal-bridge`
   packages.
 - `C-c v` is bound to `wal-org-capture-project-tasks`.
+- While editing this config, the swimming whale `posframe` is
+  displayed again for project buffers. Can be disabled by setting
+  `wal-config-show-whale-animation` to `nil`.
 
 ### Fixed